+++ wire-ios/Wire-iOS/Sources/UserInterface/SwitchBackend/SwitchBackendConfirmationView.swift
@@ -86,10 +86,6 @@ struct SwitchBackendConfirmationView: View {
                 )
             }
         }
-
-        // TODO: When iOS 16.4 is min version
-        // add this to scroll view to enable scrolling only when needed
-        // .scrollBounceBehavior(.basedOnSize, axes: [.vertical])
     }
 
     @ViewBuilder
@@ -154,163 +150,4 @@ struct SwitchBackendConfirmationView: View {
             didConfirm: { _ in }
         )
     )
-}
-
-struct PrimaryButton: SwiftUI.ButtonStyle {
-
-    func makeBody(configuration: Configuration) -> some View {
-            configuration.label
-                .padding()
-                .expandToFill(axis: .horizontal)
-                .background(Color.primaryButtonBackground)
-                .foregroundStyle(Color.primaryButtonText)
-                .clipShape(.rect(cornerRadius: 16))
-        }
-
-}
-
-struct SecondaryButton: SwiftUI.ButtonStyle {
-
-    func makeBody(configuration: Configuration) -> some View {
-            configuration.label
-                .padding()
-                .expandToFill(axis: .horizontal)
-                .background(Color.secondaryButtonBackground)
-                .foregroundStyle(Color.secondaryButtonText)
-                .clipShape(.rect(cornerRadius: 16))
-                .overlay(
-                    RoundedRectangle(cornerRadius: 16)
-                        .stroke(Color.secondaryButtonBorder, lineWidth: 1)
-                )
-        }
-
-}
-
-struct FrameExpansion: ViewModifier {
-
-    let axis: Axis
-
-    func body(content: Content) -> some View {
-        switch axis {
-        case .horizontal:
-            return content.frame(maxWidth: .infinity)
-
-        case .vertical:
-            return content.frame(maxHeight: .infinity)
-        }
-    }
-
-    enum Axis {
-
-        case horizontal
-        case vertical
-
-    }
-
-}
-
-extension View {
-
-    func expandToFill(axis: FrameExpansion.Axis) -> some View {
-        modifier(FrameExpansion(axis: axis))
-    }
-
-}
-
-extension Color {
-
-    static let primaryText = Color(uiColor: UIColor(
-            light: Asset.Colors.black,
-            dark: Asset.Colors.white
-        )
-    )
-
-    static let secondaryText = Color(uiColor: UIColor(
-            light: Asset.Colors.gray70,
-            dark: Asset.Colors.gray30
-        )
-    )
-
-    static let primaryButtonBackground = Color(uiColor: UIColor(
-            light: Asset.Colors.blue500Light,
-            dark: Asset.Colors.blue500Dark
-        )
-    )
-
-    static let primaryButtonText = Color(uiColor: UIColor(
-            light: Asset.Colors.white,
-            dark: Asset.Colors.black
-        )
-    )
-
-    static let secondaryButtonBackground = Color(uiColor: UIColor(
-            light: Asset.Colors.white,
-            dark: Asset.Colors.gray95
-        )
-    )
-
-    static let secondaryButtonBorder = Color(uiColor: UIColor(
-            light: Asset.Colors.gray40,
-            dark: Asset.Colors.gray80
-        )
-    )
-
-    static let secondaryButtonText = Color(uiColor: UIColor(
-            light: Asset.Colors.black,
-            dark: Asset.Colors.white
-        )
-    )
-
-
-}
-
-enum WireTextStyle {
-
-    case h1
-    case h2
-    case h3
-    case h4
-    case h5
-    case body1
-    case body2
-    case body3
-    case subline1
-    case link
-    case buttonSmall
-    case buttonBig
-
-}
-
-extension Text {
-
-    @ViewBuilder
-    func textStyle(_ textStyle: WireTextStyle) -> some View {
-        switch textStyle {
-        case .h1:
-            font(.title3)
-        case .h2:
-            font(.title3).bold()
-        case .h3:
-            font(.headline)
-        case .h4:
-            font(.subheadline)
-        case .h5:
-            font(.footnote)
-        case .body1:
-            font(.body)
-        case .body2:
-            font(.callout).fontWeight(.semibold)
-        case .body3:
-            font(.callout).bold()
-        case .subline1:
-            font(.caption)
-        case .link:
-            font(.body).underline()
-        case .buttonSmall:
-            fatalError("not implemented")
-        case .buttonBig:
-            font(.title3).fontWeight(.semibold)
-        }
-    }
-
 }
